@@ -23,35 +23,30 @@
     "access": "public"
   },
   "files": [
-    ".eslintrc.node.js",
-    ".eslintrc.react.js"
+    "index.js",
+    "index.d.ts"
   ],
   "main": "index.js",
+  "types": "index.d.ts",
   "scripts": {
-    "format": "prettier --write \"*.js\""
+    "format": "prettier --write \"*.js\"",
+    "lint-inspector": "npx @eslint/config-inspector"
   },
   "lint-staged": {
     "*.js": "prettier --write"
   },
   "peerDependencies": {
-    "eslint": "^8.48.0"
+    "eslint": "^9.9.0"
   },
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.11.0",
-    "@typescript-eslint/parser": "^7.11.0",
-    "eslint": "^8.50.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^18.0.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-promise": "^6.2.0",
-    "eslint-plugin-react": "^7.34.2",
+    "@antfu/eslint-config": "^2.25.1",
+    "@eslint-react/eslint-plugin": "^1.10.0",
+    "eslint": "^9.9.0",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-simple-import-sort": "^12.1.0"
+    "eslint-plugin-react-refresh": "^0.4.9",
+    "eslint-plugin-simple-import-sort": "^12.1.1"
   },
   "devDependencies": {
     "@siberiacancode/prettier": "*"
